fix: user_id type

+++ schema.sql
@@ -4,7 +4,7 @@ USE ifsales;
 
 DROP TABLE IF EXISTS users;
 CREATE TABLE users (
-    user_id INT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT AUTO_INCREMENT PRIMARY KEY,
     email VARCHAR(100) NOT NULL,
     password VARCHAR(255) NOT NULL
 );
